backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.3
celery==5.3.6
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
langchain==0.1.16
langchain-groq==0.1.3
langchain-openai==0.1.3
groq==0.5.0
slack-sdk==3.27.1
aiosmtplib==3.0.1
boto3==1.34.69
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
python-dotenv==1.0.1
